feat: introduce Messenger (#54)
This interface provides a way to both send and receive `Message`s and is
a direct replacement for `Conference.message()` and
`MessageReceivedConferenceEvent`.

It also supports direct messaging via `Messenger.send()` that accepts a
participant ID.

`MessageNotSendException` should enable clients to provide retries since
it includes both an instance of a `Message` that would've been received
if the call was successful and an optional `cause`.

Note that `MessageListener` will not be invoked for messages that are
sent by the caller (which differs from previous version). Clients are
encouraged to inject an instance of the `Message` obtained from
`SendCallback.onSuccess()` to the list they maintain manually. Previous
behavior might be undesirable for applications that would like to
utilize message types other than "text/plain" (i.e. non-chat use cases).