fix: insert black frame when VideoCapturer stops
When a client stops capturing, last sent frame will be left lingering in
the conference. While this is not an issue if this change is signalled
via REST API, it affects users in gateway calls where removing from mix
is not supported.

This commit inserts a black frame when `onCapturerStopped()` in invoked.