fix(media): switch Executors with CoroutinesScopes
This should additionally fix a rather rare, but still annoying crash
when a `Renderer` is being removed, but the underlying `VideoTrack` is
was already disposed.

Unlike other methods in `MediaStreamTrack`, `removeRenderer` does not check if
the `VideoTrack` still exists and attempts to remove and free
`VideoSink` through JNI, which happily blows up (after all, the
`VideoTrack` has already been disposed).