fix(media): switch Executors with CoroutineScopes (#170)
This should additionally fix a rather rare, but still annoying crash
when a `Renderer` is being removed, but the underlying `VideoTrack` was
already disposed.

Unlike other methods in `MediaStreamTrack`,
[`removeSink`](https://webrtc.googlesource.com/src/+/refs/heads/main/sdk/android/api/org/webrtc/VideoTrack.java#49)
does not check if the `VideoTrack` still exists and attempts to remove
and free `VideoSink` through JNI, which happily blows up (after all, the
`VideoTrack` has already been disposed).