refactor: merge base modules with -infinity modules (#251)
In a similiar fashion to what we previously did with `-coroutines`
modules, we now merge `-infinity` modules with their respective bases.
This should allow us to re-use most classes across higher level
constructs, as well as to converge on a single module that deals with
Infinity-related code - `:sdk-infinity`, which has now becomes the base.