fix(conference): certain components missing SSE
This was most obvious with the recent addition of `Layout`. The SSE would arrive before the `Theme.layout` had a chance to subscribe and the current "state" would end up being ignored.

This PR forces the underlying `event` to wait until all expected subscribers are active to initiate the connection.

An alternative implementation that relied on debouncing was considered, but was dismissed due to being unreliable (hard to guess the correct timeout value).