feat(infinity): introduce Node
This PR adds a `Node` that represents the address of the Infinity node
that the SDK will communicate with.

Note that while the `Node` includes `port`, in real world use it will
almost exclusively be 443 since Infinity forces you to use TLS over said
port. It's exposed here to allow testing against `MockWebServer`.

We also added a suspending `NodeResolver` that will return a `Node`
list and deprecated the one found in `:sdk-api` module.