refactor(api)!: simplify TokenStore (#406)
This PR replaces `AtomicReference` as the in-memory store of the `Token`
to a `MutableStateFlow`, giving us a way to observe the changes in the
`TokenStore`.

Notably, it's now an internal API (and was never meant to be used
outside of the project anyway). Consecutively, it's no longer an
interface, but a concrete class.