ci: force snyk to use runtime classpath (#412)
We're mostly interested in checking the what we include as a dependency
in the released artifacts, a bit less so when it comes to tooling.

Note the switch to `snyk-linux` instead of `snyk` - it appears that
there's an `eval` call that strips the first `'` symbol from the regex,
causing the command to fail. Invoking `snyk-linux` directly does the
trick, but will cause an error if we were to migrate to a macOS runner.