refactor(media): do not use OS level microphone mute (#514)
It appears that at least on some devices we cannot mute when the app is
running under a Work Profile.

This switches the implementation to a more reliable one based on
`AudioDeviceModule.setMicrophoneMute(Boolean)` that directly "mutes" the
underlying `AudioRecord` by injecting empty bytes into the byte buffer
instead of calling `AudioManager.setMicrophoneMuted(Boolean)`.

Note that with this change muting the microphone will no longer alter
the status in the status bar and extra steps need to be taken by the
consumers if they need this functionality.