refactor(api): switch to polymorphic serialization for Event
Previously, we'd use `when (type)` to distinguish between the type of
the incoming event and handle the deserialization manually.

With this change, we switch to generated serializers by creating a
"fake" JSON that wraps `type` and `data`. We use `#event` as the type
discriminator since we know that it will not be included in `data`.