feat(api): set breakout_capable to false by default (#527)
This means that consumers should explicitly opt-in to support breakouts
feature. By default, this is not supported and the behavior matches that
of a guest on Infinity v36+.

Note that on v34 and v35 if the participant is being transfered to a
breakout they may not receive `refer` event, which means that they will
stay in the main room.