fix(media): avoid skipping potential ICE candidates (#612)
ICE candidates that had the same `sdpMid` value were previously skipped
when `ContinualGatheringPolicy` was set to
`ContinualGatheringPolicy.GATHER_ONCE`(default), leading to potential
connectivity issues as the candidates may have different content in the
`sdp` itself. This potential fix removes this equality check and ensures
that all candidates are signalled to the remote client.